+++ changelog.md
@@ -4,6 +4,7 @@
  * Export `unstreamM` from`from Data.Vector.Generic`: [#70](https://github.com/haskell/vector/issues/70)
  * New functions: `unfoldrExactN` and `unfoldrExactNM`: [#140](https://github.com/haskell/vector/issues/140)
  * Added `iforM` and `iforM_`: [#262](https://github.com/haskell/vector/issues/262)
+ * Added `MonadFix` instance for boxed vectors: [#178](https://github.com/haskell/vector/issues/178)
 
 # Changes in version 0.12.1.2